Move NonLockingFileReader to tests where it used.
LicenseExpiredForm now uses NonBlockingReader instead.
