+++ src/NServiceBus.Core.Tests/Logging/RollingLoggerTests.cs
@@ -433,5 +433,19 @@
             }
         }
 
+        static class NonLockingFileReader
+        {
+            internal static string ReadAllTextWithoutLocking(string path)
+            {
+                using (var fileStream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
+                {
+                    using (var textReader = new StreamReader(fileStream))
+                    {
+                        return textReader.ReadToEnd();
+                    }
+                }
+            }
+        }
+
     }
 }
