--- src/NServiceBus.Core/Licensing/NonLockingFileReader.cs
@@ -1,18 +1,0 @@
-namespace NServiceBus
-{
-    using System.IO;
-
-    static class NonLockingFileReader
-    {
-        internal static string ReadAllTextWithoutLocking(string path)
-        {
-            using (var fileStream = new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
-            {
-                using (var textReader = new StreamReader(fileStream))
-                {
-                    return textReader.ReadToEnd();
-                }
-            }
-        }
-    }
-}
