+++ src/NServiceBus.Core/NServiceBus.Core.csproj
@@ -589,7 +589,6 @@
     <Compile Include="Licensing\AuditInvalidLicenseBehavior.cs" />
     <Compile Include="Licensing\LicenseReminder.cs" />
     <Compile Include="Licensing\LicenseSources.cs" />
-    <Compile Include="Licensing\NonLockingFileReader.cs" />
     <Compile Include="Licensing\StaThreadRunner.cs" />
     <Compile Include="Licensing\LicenseExpiredFormDisplayer.cs" />
     <Compile Include="Logging\ColoredConsoleLogger.cs" />
